requirements.txt
revtorch
pytorch
tensorflow